Fix signed/unsigned warnings

+++ src/rng.hpp
@@ -67,7 +67,7 @@ template <typename T> inline void xoshiro_jump(rng_state_t<T> &state) {
   uint32_t s1 = 0;
   uint32_t s2 = 0;
   uint32_t s3 = 0;
-  for (int i = 0; i < sizeof JUMP / sizeof *JUMP; i++)
+  for (long unsigned int i = 0; i < sizeof JUMP / sizeof *JUMP; i++)
     for (int b = 0; b < 32; b++) {
       if (JUMP[i] & UINT32_C(1) << b) {
         s0 ^= state[0];
@@ -96,7 +96,7 @@ template <typename T> inline void xoshiro_long_jump(rng_state_t<T> &state) {
   uint32_t s1 = 0;
   uint32_t s2 = 0;
   uint32_t s3 = 0;
-  for (int i = 0; i < sizeof LONG_JUMP / sizeof *LONG_JUMP; i++)
+  for (long unsigned int i = 0; i < sizeof LONG_JUMP / sizeof *LONG_JUMP; i++)
     for (int b = 0; b < 32; b++) {
       if (LONG_JUMP[i] & UINT32_C(1) << b) {
         s0 ^= state[0];
